+++ opts/global.go
@@ -578,166 +578,6 @@ type AxisLine struct {
 	LineStyle *LineStyle `json:"lineStyle,omitempty"`
 }
 
-// XAxis is the option set for X axis.
-// https://echarts.apache.org/en/option.html#xAxis
-type XAxis struct {
-	// Name of axis.
-	Name string `json:"name,omitempty"`
-
-	// Location of axis name.
-	//
-	// Options:
-	// 'start'
-	// 'middle' or 'center'
-	// 'end'
-	NameLocation string `json:"nameLocation,omitempty"`
-
-	// Gap between axis name and axis line.
-	NameGap int `json:"nameGap,omitempty"`
-
-	// Type of axis.
-	// Option:
-	// * 'value': Numerical axis, suitable for continuous data.
-	// * 'category': Category axis, suitable for discrete category data.
-	//   Category data can be auto retrieved from series.data or dataset.source,
-	//   or can be specified via xAxis.data.
-	// * 'time' Time axis, suitable for continuous time series data. As compared to value axis,
-	//   it has a better formatting for time and a different tick calculation method. For example,
-	//   it decides to use month, week, day or hour for tick based on the range of span.
-	// * 'log' Log axis, suitable for log data.
-	Type string `json:"type,omitempty"`
-
-	// Set this to false to prevent the axis from showing.
-	Show types.Bool `json:"show,omitempty"`
-
-	// Category data, available in type: 'category' axis.
-	Data interface{} `json:"data,omitempty"`
-
-	// Number of segments that the axis is split into. Note that this number serves only as a
-	// recommendation, and the true segments may be adjusted based on readability.
-	// This is unavailable for category axis.
-	SplitNumber int `json:"splitNumber,omitempty"`
-
-	// It is available only in numerical axis, i.e., type: 'value'.
-	// It specifies whether not to contain zero position of axis compulsively.
-	// When it is set to be true, the axis may not contain zero position,
-	// which is useful in the scatter chart for both value axes.
-	// This configuration item is unavailable when the min and max are set.
-	Scale types.Bool `json:"scale,omitempty"`
-
-	// The minimum value of axis.
-	// It can be set to a special value 'dataMin' so that the minimum value on this axis is set to be the minimum label.
-	// It will be automatically computed to make sure axis tick is equally distributed when not set.
-	Min interface{} `json:"min,omitempty"`
-
-	// The maximum value of axis.
-	// It can be set to a special value 'dataMax' so that the minimum value on this axis is set to be the maximum label.
-	// It will be automatically computed to make sure axis tick is equally distributed when not set.
-	Max interface{} `json:"max,omitempty"`
-
-	// Minimum gap between split lines. For 'time' axis, MinInterval is in unit of milliseconds.
-	MinInterval float64 `json:"minInterval,omitempty"`
-
-	// Maximum gap between split lines. For 'time' axis, MaxInterval is in unit of milliseconds.
-	MaxInterval float64 `json:"maxInterval,omitempty"`
-
-	// The index of grid which the x axis belongs to. Defaults to be in the first grid.
-	// default 0
-	GridIndex int `json:"gridIndex,omitempty"`
-
-	// Split area of X axis in grid area.
-	SplitArea *SplitArea `json:"splitArea,omitempty"`
-
-	// Split line of X axis in grid area.
-	SplitLine *SplitLine `json:"splitLine,omitempty"`
-
-	// Settings related to axis label.
-	AxisLabel *AxisLabel `json:"axisLabel,omitempty"`
-
-	// Settings related to axis tick.
-	AxisTick *AxisTick `json:"axisTick,omitempty"`
-
-	// Settings related to axis pointer.
-	AxisPointer *AxisPointer `json:"axisPointer,omitempty"`
-}
-
-// YAxis is the option set for Y axis.
-// https://echarts.apache.org/en/option.html#yAxis
-type YAxis struct {
-	// Name of axis.
-	Name string `json:"name,omitempty"`
-
-	// Location of axis name.
-	//
-	// Options:
-	// 'start'
-	// 'middle' or 'center'
-	// 'end'
-	NameLocation string `json:"nameLocation,omitempty"`
-
-	// Gap between axis name and axis line.
-	NameGap int `json:"nameGap,omitempty"`
-
-	// Type of axis.
-	// Option:
-	// * 'value': Numerical axis, suitable for continuous data.
-	// * 'category': Category axis, suitable for discrete category data.
-	//   Category data can be auto retrieved from series.data or dataset.source,
-	//   or can be specified via xAxis.data.
-	// * 'time' Time axis, suitable for continuous time series data. As compared to value axis,
-	//   it has a better formatting for time and a different tick calculation method. For example,
-	//   it decides to use month, week, day or hour for tick based on the range of span.
-	// * 'log' Log axis, suitable for log data.
-	Type string `json:"type,omitempty"`
-
-	// Set this to false to prevent the axis from showing.
-	Show types.Bool `json:"show,omitempty"`
-
-	// Category data, available in type: 'category' axis.
-	Data interface{} `json:"data,omitempty"`
-
-	// Number of segments that the axis is split into. Note that this number serves only as a
-	// recommendation, and the true segments may be adjusted based on readability.
-	// This is unavailable for category axis.
-	SplitNumber int `json:"splitNumber,omitempty"`
-
-	// It is available only in numerical axis, i.e., type: 'value'.
-	// It specifies whether not to contain zero position of axis compulsively.
-	// When it is set to be true, the axis may not contain zero position,
-	// which is useful in the scatter chart for both value axes.
-	// This configuration item is unavailable when the min and max are set.
-	Scale types.Bool `json:"scale,omitempty"`
-
-	// The minimum value of axis.
-	// It can be set to a special value 'dataMin' so that the minimum value on this axis is set to be the minimum label.
-	// It will be automatically computed to make sure axis tick is equally distributed when not set.
-	Min interface{} `json:"min,omitempty"`
-
-	// The maximum value of axis.
-	// It can be set to a special value 'dataMax' so that the minimum value on this axis is set to be the maximum label.
-	// It will be automatically computed to make sure axis tick is equally distributed when not set.
-	Max interface{} `json:"max,omitempty"`
-
-	// The index of grid which the Y axis belongs to. Defaults to be in the first grid.
-	// default 0
-	GridIndex int `json:"gridIndex,omitempty"`
-
-	// Split area of Y axis in grid area.
-	SplitArea *SplitArea `json:"splitArea,omitempty"`
-
-	// Split line of Y axis in grid area.
-	SplitLine *SplitLine `json:"splitLine,omitempty"`
-
-	// Settings related to axis label.
-	AxisLabel *AxisLabel `json:"axisLabel,omitempty"`
-
-	// Settings related to axis line.
-	AxisLine *AxisLine `json:"axisLine,omitempty"`
-
-	// Settings related to axis pointer.
-	AxisPointer *AxisPointer `json:"axisPointer,omitempty"`
-}
-
 // TextStyle is the option set for a text style component.
 type TextStyle struct {
 	// Font color
@@ -1272,16 +1112,6 @@ type Calendar struct {
 	Silent types.Bool `json:"silent,omitempty"`
 }
 
-// Polar Bar options
-type Polar struct {
-	ID      string    `json:"id,omitempty"`
-	Zlevel  int       `json:"zlevel,omitempty"`
-	Z       int       `json:"z,omitempty"`
-	Center  [2]string `json:"center,omitempty"`
-	Radius  [2]string `json:"radius,omitempty"`
-	Tooltip Tooltip   `json:"tooltip,omitempty"`
-}
-
 type PolarAxisBase struct {
 	ID           string     `json:"id,omitempty"`
 	PolarIndex   int        `json:"polarIndex,omitempty"`
@@ -1298,21 +1128,6 @@ type PolarAxisBase struct {
 	LogBase      float64    `json:"logBase,omitempty"`
 	Silent       types.Bool `json:"silent,omitempty"`
 	TriggerEvent types.Bool `json:"triggerEvent,omitempty"`
-}
-
-type AngleAxis struct {
-	PolarAxisBase
-	Clockwise types.Bool `json:"clockwise,omitempty"`
-}
-
-type RadiusAxis struct {
-	PolarAxisBase
-	Name          string     `json:"name,omitempty"`
-	NameLocation  string     `json:"nameLocation,omitempty"`
-	NameTextStyle TextStyle  `json:"nameTextStyle,omitempty"`
-	NameGap       float64    `json:"nameGap,omitempty"`
-	NameRadius    float64    `json:"nameRotate,omitempty"`
-	Inverse       types.Bool `json:"inverse,omitempty"`
 }
 
 var newlineTabPat = regexp.MustCompile(`\n|\t`)
@@ -1568,30 +1383,6 @@ type ViewControl struct {
 	// Rotate Speed, (angle/s).
 	// default 10
 	AutoRotateSpeed float32 `json:"autoRotateSpeed,omitempty"`
-}
-
-// Grid Drawing grid in rectangular coordinate.
-// https://echarts.apache.org/en/option.html#grid
-type Grid struct {
-	// Distance between grid component and the left side of the container.
-	Left string `json:"left,omitempty"`
-
-	// Distance between grid component and the right side of the container.
-	Right string `json:"right,omitempty"`
-
-	// Distance between grid component and the top side of the container.
-	Top string `json:"top,omitempty"`
-
-	// Distance between grid component and the bottom side of the container.
-	Bottom string `json:"bottom,omitempty"`
-
-	// Width of grid component.
-	Width string `json:"width,omitempty"`
-
-	ContainLabel types.Bool `json:"containLabel,omitempty"`
-
-	// Height of grid component. Adaptive by default.
-	Height string `json:"height,omitempty"`
 }
 
 // Dataset brings convenience in data management separated with styles and enables data reuse by different series.
